fix(treesitter)!: correctly handle query quantifiers
Query patterns can contain quantifiers (e.g. `(foo)+ @bar)`, so a single
capture can map to multiple nodes. The iter_matches API can not handle
this situation because the match table incorrectly maps capture indices
to a single node instead of to an array of nodes.

Update the match table to map capture indices to an array of nodes.

BREAKING CHANGE: The match returned by Query:iter_matches() is now a map
from capture_id to a *list* of nodes instead of map from capture_id to
*single* node. Predicates and directives must also be updated.

Co-authored-by: Gregory Anders <greg@gpanders.com>